Kerem/hnt 3972 use aws secret manager for rds master credentials (#4470)
we're no longer using rds-iam for authenticating the root user, and
instead using the aws secrets manager. the rds-iam user strategy assumed
that the master user was configured to work with iam, which can only
happen if you run an init sql query to make it so. this worked with
transient envs + single node since we could just log in as master once,
enable iam auth, and then have the transient envs automatically obtain
iam auth by virtue of cloning the test-beta database. moving forward,
new nodes with no prior history will enter the config, so we can no
longer work with the assumption that iam will be enabled for them.

tl;dr we're using secrets manager to authenticate as root user to run
the initial db configs post provision