Encrypted Usernames + Collision Detection (#4509)
This PR introduces encrypted usernames in the SDK only — not available
in the client yet.

Usernames are pretty logic heavy — let's put them in their own class
(`Usernames`). Display names will also grow in complexity once we add
deduplication and we could do the same thing for those. Much easier to
test.

### Changes
- `displayNames.test.ts` -> `userMetadata.test.ts`
- sort the events in the snapshot based on `eventNum`.
- on submit, the plaintext username is hashed using
`usernameChecksum(username: string, streamId: string)`
- `checksum` is an optional property added to `EncryptedData` (LMK if
you can think of a better place to put it!)
- if the checksum is missing, we will refuse to show the username in the
client
- if the checksum doesn't match on decrypt, we will refuse to show the
username in the client
- if a client _changes_ their username, the old username is instantly
available for other clients to use
- when we add usernames to the client, let's expose `usernameAvailable`
to the app, and make the `setUsername` function throw an error if
there's a collision

It's likely that the hook for fetching user data will want something
similar to:
```
{
  username: string
  usernamePending: boolean
  displayName: string
  displayNamePending: boolean
}
```
which will be easy to set up with this approach

stacked on #4482