app: wallet link on public town page (#4490)
On the public town page:
- always shows "Join Town" after logged in
- app performs entitled check before join tx. If you are entitled, you
mint and join
- if you are not entitled, a modal will pop up that informs you which
assets you need to join
- entitlements are checked on any newly linked wallet. Additionally, if
an asset has not yet been confirmed, each linked wallet + root wallet is
polled for the asset. So if the user sends the asset to any of their
wallets, the UI will update

<img width="1206" alt="Screenshot 2023-12-11 at 12 27 25 PM"
src="https://github.com/HereNotThere/harmony/assets/15717901/d5bfa3e4-32fa-4b20-9de3-e1d63cf542e3">

Once the transient env is up the above town can be previewed at:
https://4490.app-preview.towns.com/t/SPCE-d4_HnvnJQLfRGqLVEGZtN/?invite

One big caveat: The client checks you are entitled to the space against
all of your linked wallets. But calling `joinTown` on the contract
results in an entitlement failure. This results in a situation where you
can have all the required assets in a linked wallet, which the UI shows,
but if you actually try to join then it will fail. That will need to be
addressed elsewhere, meanwhile this update doesn't effect joining a
non-gated town, nor a town whose asset resides in your privy wallet, so
this should be fine for now.