Cache GetStream() requests (#4458)
We can improve app performance and decrease the load on the nodes a bit
here by checking if we have the stream in the cache.

My working hypothesis is that we will need to check with
`getLastMiniblockHash` to make sure that our cached data isn't stale. If
it is, we'll need to refetch the stream from the nodes using
`GetStream()`. Maybe this isn't enough? We may need the last event hash?

Stacked on #4457