introducing new labels for transient single node + multi node (#4567)
Multi-node infra work is done:
https://github.com/HereNotThere/harmony/pull/4558


![image](https://github.com/HereNotThere/harmony/assets/5973785/2f41a1ce-45eb-4a8e-9b36-47bddace6774)

Now, we need to change our labels one more time to streamline the
developer experience. Moving forward, we will retire `provision full
transient environment` in favor of `single-node transient environment`
and `multi-node transient environment`.

If `single-node`, numNodes is set to 1. If `multi-node`, numNodes is set
to 10. If no label exists, numNodes is set to 0. This also implies that
we can remove the `lite` language from the workflow. Everything reduces
down to the number of nodes, including whether we'll run docker builds,
continuous deploys or synthetic tests.