source step added to carry env vars (#4601)
I found another transient-env <> river node connection bug. Basically
this script sets the env vars and then runs the build. But at the time
of running this script for the first time, the env vars aren't set, so
the first run builds against test-beta. A subsequent run will work with
the correct environments. However, if you access the website before the
second run, your service worker will keep talking to test-beta. This is
why it's crucial to set the node url on the first try.

This PR makes the first step export its environment variables via
`source`, so that the build step, even on the first run, understands the
new csb homserver url