controlling clean vs cloned transient dbs via pr labels (#4622)
pr labels can now control whether the transient env is with a clean db
or a cloned db. this is due to how certain workflows actually require us
to start with a clean db. for example, multinode currently needs a clean
db, but eventually it will need to work from a test-beta cloned db.
another example is the upcoming sepolia migration which will require a
clean db while testing in a transient environment. since clean
environments are going to be a lot less frequent, I'm color-coding them
in `red` to indicate that you're about to do something unusual. Lmk what
you guys think. If approved, I will update the notion doc to reflect the
new labels


![image](https://github.com/HereNotThere/harmony/assets/5973785/4eb82fd9-c191-47ac-a1a3-0e2769639f76)