app: automatic z-index stacking context (hnt-4067) (#4633)
(I'm probably going to rename and refactor a bit tomorrow, but the main
idea is there)

problem: even if z-index should be avoided we needed it for some
specific places where the dom order conflicts with the layer order.
We're using some presets (e.g. `tooltips`) to keep things clean. However
when using panels on mobile this doesn't really work: each new panel
that animates in needs to be higher than z-indexed elements, and
everything that comes after gets flattened, it works to some extend but
not quiet.

solution: we can add a layer (or stacking?) context each time we want to
add a new panel. This way we don't have to hack around or modify our
current semantics. Every new stack increments the z-index by 100, then
we can easily offset each of our existing presets with `layer * 100 + z`
- This all happens under the hood as long as we add a new context.

- root - layer 0
  - `@ui` (z-index: 10)
  - `@tooltips`: (z-index: 20) 
- panel 1 - layer 1 
  - `@ui` (z-index: 110)
  - `@tooltips`: (z-index: 120)
- nested panel - layer 2
  - `@ui` (z-index: 210)
  - `@tooltips`: (z-index: 220)  
  ...