app: tidy upp client unread markers (hnt-4183) (#4701)
this PR simplifies client-side mark-as-unread mechanism and fixes some
edge-cases. There's been an issue since I started persist the "new"
marker on the screen. Separating logic between "visual delimiter" and
"mark-as-unread" make things more predictable.

- `<NewDivider>` and `<DateDivider>` are only visual markers and won't
trigger receipts
- An invisible `<FullyReadObserver>` is inserted besides the event that
needs to enter the viewport
- the persisted visual marker will only change when entering a channel
or if a message is received when the browser was blurred