app: safari - fix central area remaining blank until scrolling the sidebar (#4712)
safari still acts strangely when it comes to layer composition.

I can repro the bug here by going to the mentions tab, observing a blank
screen, then scrolling the left hand sidebar after which the central
area reveals. There's nothing related in the code, even removing the
contents of the sidebar still triggers the glitch so unless it's related
to `Allotments` it's just pretty random.