lib: restore hasTownMembership check when trying to join town (#4735)
When you leave a town, you can't rejoin b/c the contract reverts with a
new reason.

We used to check membership before calling the joinTown mint fn. It got
removed [here](https://github.com/HereNotThere/harmony/pull/4442). Also
part of that PR, we catch the specific `Membership__AlreadyMember`
reversion, and continue you to join the town in river. The issue now is,
a member who already has the NFT, the contract reverts w/
`BalanceQueryForZeroAddress` - for an unknown reason, which I've talked
with @giuseppecrj about. So you never rejoin.

This is a quick fix so we just rejoin you if you already have the
membership.