fix start_dev.sh to remove postgres container on kill-session, and disable the create_multi.sh (#4786)
3 changes:
1. Move "fund wallet" after `run_single.sh`. On a new machine, or
cleaned up dirs, the `run_files/` does not exist until `run_single.sh`
is run.
2. When i run the `app` after `start_dev`, creating the user stream
fails because the DB already has the stream, but on chain, the stream
does not exists since it's a new anvil instance. It is cleaner to
remove the postgres container after `kill-session`. That way, when
re-running `start_dev`, everything starts cleanly. This is the same
behavior as tasks.json `Start Local Dev` and `KillAllLocalDev`.
3. The `create_multi.sh` fails with error described in
[hnt-4317](https://linear.app/hnt-labs/issue/HNT-4317/create-multish-call-in-the-start-devsh-fails).
Disabledfor now.