userops: add skandha bundler, test setup (#4809)
I've added a simple test as a sanity check that can be run from the
casablanca/web3 dir with `yarn test:funds`. This test sends a userop to
initiate a transfer of funds.
Using the eth-infinitism bundler, this test fails.
Using the skandha bundler (which I've added under /servers), this test
passes.

The tests in `userops-space-dapp-transactions` pass only when
`.env.test` vars are pointed to stackup urls. When trying to run against
skandha bundler, the tests fail.