Memoize RichTextEditorWithoutBoundary (#4830)
When I enabled react profiling and turned off name mangling, I saw this
component was identified as the trigger for re-rendering when the app
was idle (I believe in DM scenarios). I saw the other instance was
wrapped in a React.memo, so I put one in, and the re-render because of
this stopped. I didn't investigate further. I did find another trigger
as well, Theo put up a fix, for a missing dependency array in a
useEffect.