Move the textfield memory out of useStore (#4831)
I did an experiment and put `console.log(!!!!! main)` in Main.tsx
It fires every time we type a letter in the input box!!!
It stops firing if I comment out `<RememberInputPlugin
storageId={props.storageId} />`

This pluging was setting a property in useStore

Everywhere that was listening to a partial of useStore was then
re-running their hooks, incuding useRootTheme, which is in Main

I expect there is a way to get quieter partial updates out of a zustand
store? But breaking it out seemed like the easier path forward.

Now the app no longer does a full render calculation when you type one
letter.