using a single base chain network url secret instead of N secrets (#4870)
Removing the N nodes network url configuration. this is legacy code that
we no longer require. it's also a stepping stone for anvil forks of base
chain, as anvil will require a network url, and it makes little sense to
use the lambdas to create another secret - or reuse a node's own network
url secret. now, we just have one secret in the global-shared
environment, and both the nodes and the fork will reuse it.

---------

Co-authored-by: mattyait <mattyait@gmail.com>