userops: refactor UserOpSpaceDapp to use paymaster proxy (#4880)
For now this middleware is defined in UserOpSpaceDapp. But I will do
another PR for hoisting it to the app level as a passed callback.

userop.js - we have to pass the paymaster middleware in the
simpleAccount.init, we can't use useMiddleware. I added a middleware
function that tries to call the paymaster proxy. If that fails, it falls
back to estimating the gas the way userop does it when no paymaster
middleware is present. Test cases work in these scenarios:

- userops-sanity.test.ts - running anvil locally, no paymasterProxyUrl.
Verifies that userop will still pass if paymaster fails, and user has
their own funds
- userops-space-dapp-transactions.test.ts - setting env vars to use
local paymaster proxy, and sepolia/stackup node. Verifies the paymaster
proxy can fund the userop when working with sepolia/stackup (createTown)

@jterzis had to make a small change to stackup worker response, I'll
redeploy the worker before merging