Remove Snapshot requirement in Stream+StreamStateView initializers (#4889)
This PR removes the `Snapshot` requirement in the `Stream` +
`StreamStateView` initializers.
Part of the local storage project, which requires us to be able to
initialize streams before we have access to the stream content. We can
later prioritize the order in which we load streams, either from disk or
from `GetStream`.