temp hotfix: always force fresh GetStream (#4931)
We're making a call to `syncStreams` from `createSyncLoop` in
`syncedStreams.ts`.
In the call, we're including all the sync cookies we have. If one of
these cookies is invalid, the call fails and no streams are synced. We
missed taking this into account last night. This temporary hotfix will
make sure that we always download fresh streams.

We'll need to discuss how to solve this. Perhaps return a list of failed
stream ids from `syncStreams`. We could rely only on `addStreamToSync`
for syncing streams, but that can break when the sync loop restarts.