SDK: set high priority ids in client init (#4945)
This PR makes sure that we set high priority stream ids when
initializing the SDK. That way no other streams can jump the queue. This
means that we'll need to catch these ids as soon as possible after
initializing the app.

We don't yet have bookmark support for GDMs/DMs, but when we do, let's
add support for prioritizing those as well.