lib: add userop integration tests for createSpace and joinTown (#4949)
when creating or joining a town, the membership nft is minted to the
smart account. we need to link the smart account address to the root
(privy) address so that entitlement checks pass and the user can
actually join

this PR updates the user ops for createSpace and joinTown to batch
execute a wallet link of the smart account, if link does not previously
exist. since either of these operations are precursors to any other
transactions in the app, i think it's safe to link the wallet here,
instead of on account creation. it creates new integration/userops test
group in lib

heads up @jterzis i updated the stackup worker to accommodate this flow,
will deploy before merging.
I'll follow up with @mechanical-turk about getting these tests running
in CI.