SDK: only return miniblocks from persistence store if all of them are found (#5020)
Finally tracked down the issue that had been bugging me. I was able to
replicate the missing messages error on my desktop and it turned out
that the issue was related to how miniblocks were returned from the
persistence store.

When inspecting IndexedDB, I noticed that my client had a gap among the
miniblocks. The persistence store would happily return the miniblocks it
found, instead of returning `[]` on an incomplete fetch.

Still investigating how the gap appeared in the first place, but this
fix is something we'll need either way.