Fix for out of order messages when a redactiom gets confirmed (#5047)
tldr Things got complicated, complicated things can have bugs.

An event is updated multiple times

appended - usually an encrypted event
decrpyted (update) - swap out the encrypted content with decrypted
content
confirmed (update) - when a miniblock comes in, update the confirmed
event num

if the decrypted content includes a redaction, we replace the redacted
event content, then append a “redaction action” to the timeline.

If an event arrived in between decrypting and confirming, then the
redaction action would get pulled out of the timeline and appended on
the end, resulting in out of order messages. This shouldn’t be visible
because we don’t render the redaction actions, but it was wrong and
could cause other issues with unread bookkeeping and its just confusing.

fixed.