lib: remove wagmi (#5014)
We don't really use wagmi for much in `lib` anymore, and its presence is
confusing. Having wagmi made a lot more sense when towns was
accommodating multiple blockchain accounts per user.

But now the lib no longer manage or knows about the signer, it just
passes it on to space dapp for writes. Using wagmi `accounts` gave us a
backdoor into the connected account, but we're not really using it for
anything. And instead of transforming a viem `publicClient` to ethers,
we can just instantiate a ethers provider and pass it on to the
spaceDapp.

Instead of passing a `chainId` to ZionContextProvider, it now accepts a
viem `Chain`, from which the chainId and rpcUrls are derived. We can
copy and define this interface ourselves later if we want to remove viem
as well.