SDK: start syncStreams once all streams have loaded (#5050)
By deferring the call to `SyncStreams` until _after_ we've initialized
all the streams, we will save 1 call per stream. In my case ~80 calls to
`AddStreamToSync`.

We're now in a place where we could simply remove the requirement of
calling `startSync` and let the client handle it, but we have some tests
that depend on being able to defer startSync so I suggest that we keep
it for now.

Also removes the `forceDownload` functionality we had before — it's no
longer needed.