Use userStream to check my joined channels (#5106)
We already have all channels info with names in a space. The existing
solution is to wait for all channel streams created then we can build
RoomData and check if the user is a member or not.

It seems like we can directly check from userSteam.joinedSteams, which
is relatively faster to display the joined channels in the left panel,
without waiting for all channel streams created.

From user experience, previously, the joined channels were displayed one
by one, and it's very visible to the user. With this PR, the user will
see all joined channels once.