HNT-4333 Fix BAD_PREV_MINIBLOCK_HASH from server channel creation (#5115)
We’re getting lots of errors like this one in ci:
https://github.com/HereNotThere/harmony/actions/runs/7661956042/job/20883321825?pr=5111

The user is approximately 11 seconds old
She creates a space and a default channel for that space
Space creation updates the users userStream and triggers a new miniblock
But the miniblock gets created after the call to create channel
Create channel grabs the user’s stream, then does three things
- creates a new channel stream
- updates the space stream
- updates the user stream
By the time it gets around to #3, the users userStream has a new block,
but the calling code has a reference to the old lastBlockHash, which is
now not the last block, and greater than 11 seconds old.

I’ve updated the code to allow the user to add an event pointing to the
second to last block if the last block is less than 11 seconds old. The
code is actually more general, it allows adding an event pointing to
block N if block N is the last block, or block N+1 is less than 11
seconds old.

I think this is what we actually intended with recency constraints, and
will dramatically reduce the number of BAD_PREV_MINIBLOCK_HASH errors.