Add isLoadingMemberships to determine if we're loading memberships (#5114)
There is a blinking issue on app start if the user has already logged
in. This is due to we're not sure if we sync the membership or if the
user is not a member in a space, right now it seems we just wait for a
random time. This PR will provide a deterministic way for this.

Before this PR, screen transitions are:
welcome -> public town page (sometimes displayed) -> message timeline

After this PR, screen transitions are:
welcome -> message timeline