app: fix join flow for logged in users, cleanup join routes (#5149)
This cleans up some components/routing around joining/loading a space. 

https://github.com/HereNotThere/harmony/pull/5114 introduced a case
where the user could be logged in, but not joined to a town, and they
hit a wrong, outdated join flow that no longer works.

Renames isLoadingMemberships to hasLoadedMemberships.

CheckValidSpaceOrInvite was an early iteration before ValidateMembership
was introduced, and is basically moot now, so I removed it and ported
relevant tests to ValidateMembership.test. I _think_ this should not
cause bugs or add back in any screen flickers 😅