new docker image for service discovery (#5173)
Now that we're removing the alb, we need a more manual way to handle
service discovery. This PR introduces a new docker image, which we'll be
running as a sidecar to river nodes. Whenever a river node goes up, this
sidecar docker container will set the cloudflare A record to its own
public ip. If a node goes down, and then up again, this part will run
again to re-set the A record.