SDK: auto scrollback N pages if the stream has cached content (#5165)
Not sure if I like having this in the SDK, but I think it has to be if
we want to avoid re-renders on init with the current setup.

Getting a large number of tickets related to:
- content re-renders on load (client has to paginate 3-4-5 times to fill
1 screen)
- "missing threads" — threads beyond the last snapshot will not auto
load

I think overall we have an issue with key solicitation related events
overwhelming the channels. Perhaps we should move them into Spaces. Most
people in a space will be a member of most channels anyway, and we
already push requests for username decryption keys on the space.


![scrollbacks](https://github.com/HereNotThere/harmony/assets/352982/67f87e1d-4283-460c-a71a-13a0e5e060e2)