SDK: use dictionary for user membership ops (#5142)
We had a nasty bug this weekend which appeared only for @brianathere ,
where we found that he was both invited and joined to a GDM stream.

The snapshot membership state is a dictionary, let's use a dictionary
instead of different sets. Users should only be able to be in one
membership state per stream simultaneously.