Prevent streams from emitting events after they are removed from sync (#5333)
I added a listener to streamInitialized here:
https://github.com/HereNotThere/harmony/actions/runs/7825605515/job/21377821261?pr=5323

And tried to run something on every stream. It failed the media tests!
The test would finish and stop the sync before initStream was over,
which would remove the stream from the streams map, before
streamInitialized was called, and the stream wouldn’t exist.

I think this could actually happen in the wild if you joined and left a
channel really fast maybe?

I think this api makes sense, call stop on the stream when removing it
from sync.