Stream Hardening: Isolate rules for add event (#5283)
Made a builder pattern to make the rules. Now we can actually read them
and know whats going on.

Flipped the order that we submit membership events, now all memberships
should go through the user stream. The rest should be derived events.

The idea is that you 
submit a user membership join for channelA
the code checks that you can submit that event, then tells the service
that you need to submit an event to the channel stream
the service goes back up to the forwarder, submits a channel event
if the event causes no change, the call should succeed, if it can be
added, it will succeed, otherwise it will fail
the channel membership add requires a space membership, so that goes to
the space stream
if the space event exists, the request unwinds, the channel event is
added
it unwinds again, the user membership event is added

This is all repeatable. You can keep calling this until it succeeds and
your user membership is added. At that point you are done.

I still need to fix lots of tests to join first