lib: integration test fixes for when anvil has a BLOCK_TIME (#5408)
These are the changes i made to get lib integration tests to pass when
there's a block time.

We have to retry space creation b/c ~~of reentrancy guard~~ of unknown
reasons. The tx fails, but its not clear why.

I shifted a couple `tx.wait` to `zionclient.waitForXxx` just because
that's how we do it in the actual app and so maybe help prevent missing
something.

EDIT: actually, I'm not certain that this is caused by the reentrancy
guard. I tried to reproduce w/in the app, and I do get a clear revert
reason, unlike the empty error message we've encountered in CI. I think
this may be because of the `_deploy` call w/in space creation.
@giuseppecrj has mentioned that BLOCK_TIME can break deployments.
Regardless, retrying space creation seems to eventually work.

Believe this could be related
https://github.com/foundry-rs/foundry/issues/4843. Added some code to
try to get a better error message