SDK: add minimal worker queue in syncedStreams (#5348)
We have a bottleneck in `syncedStreams.ts`. Depending on the response,
we could be performing a large amount of work in the callback from
`syncStreams`, which means that the next sync item sent from the node
will have to wait until the client is done processing it.

Let's offload incoming sync updates to a queue and process them from
there.

`setTimeout` has a minimum timeout of 1ms. Exploring the use of
`queueMicrotask` here instead.
Obviously we will want to parallelize the work in the tick. However, we
need to be smart about it. `Promise.all(queue)` could have all kinds of
bad side effects since ordering isn't guaranteed, and we could end up
performing mutations on a stream in the wrong order. We could easily
parallelize items from _different_ streams though. Will be a follow-up
PR.

**We are quite slow** to update the app. When sending 50 messages at the
same time, the last received sync response is placed in the queue about
8-10s before the last item is processed.