bump connect deps (#5346)
trying to chase down the issue with some clients where the sync stalls.
the syncStreams call is active, but in reality it seems like it has
stalled. not sure if it's related to the connect version but worth a
shot.
Would we also want to bump the version in the backend?