app: add user ops to channel and role transactions (#5434)
User operations are used for channel and role transactions. The only
remaining txs not to use user operations are:
- updating space name - needs work in paymaster proxy
- linking wallet - blocked, see
https://linear.app/hnt-labs/issue/HNT-4908/refactor-wallet-linking-for-smart-accounts-or-dont-use-it-with-smart

This PR also adds the tx confirm modal to each of these transactions -
in the case that the paymaster proxy rejects, the user sees the modal w/
gas estimation.

This PR also adds a `yarn dev:userops` to app, which will boot up the
app configured for user operations. We can't run user operations against
anvil, so you have to develop against live river/base-sepolia. Basically
the same as what you can do now, with added configuration to turn on
user operations.