Tak/hnt 4972 show a default notification message to avoid getting throttled (#5511)
If we can't decrypt a DM / GDM message because the device does not have
keys yet, the notification could be swallowed. If notifications are not
shown for too many times, the device's registration may be deleted by
APNS.
Show a default message in the missing keys scenario so that notification
is almost never hidden from the user. Hopefully, the user taps the
notification to launch the app and syncs the keys so that subsequent
notifications can be decrypted.