Change streamView.forEachEvent to iterate over all events (#5581)
This ties into the change made here:
https://github.com/HereNotThere/harmony/pull/5570

The joinable stream functions were programmed with forEachEvent and as
such ignored the events in the minipool. This turned out to be not
useful and caused race conditions when trying to validate adding events
and creating streams.

Serge and I talked about renaming this to `forEachMiniblockEvent`

But, turns out there’s nowhere in the code where we don’t want to
iterate over all the events. So I took the liberty to fix the behavior
and remove all the `.minipool.forEachEvent` calls. I think this is what
we want.