Normalize packages/generated dir structure, build river types (#5579)
Closes HNT-5012/HNT-5033 - This PR normalizes packages/generated
directory structure to reflect the various environments that our
contracts are deployed in (dev reflecting localhost, v3 reflecting
remote environment) removing the conditionality on chainId, which varies
depending on not just environment, but also contract.

Though I kept it in this PR and added a line for each (address, chainId)
tuple, it turns out we don't use .chainId file generated as part of
contract deployment to identify the app or node runtime chainId. For the
River stream node, the chainId's needed are passed as config via
environment variables for base, river chain, respectively at runtime.
For the app, we only connect to Base and that chainId is also passed in
as config to the render app. We'll likely need to add a new var to the
app to support the forthcoming `RiverChainDapp`.

The initial inspiration for this simpler directory structures came from
@sergekh2 who started employing this pattern in the River node
deployment scripts. Below is a tree of before and after this PR what
`packages/generated` looks like.

Secondly, this PR adds typings for river chain contracts, which will be
needed subsequently to develop a Shim in the app to support Node
resolution.


Before this PR:
```
./packages/generated
├── README.md
├── base_sepolia
│   ├── addresses
│   └── v3
├── localhost
│   ├── addresses
│   └── v3
├── package.json
└── river
    └── addresses
```
After this PR:
```
./packages/generated
├── README.md
├── v3 (base_sepolia types, addresses)
│   ├── addresses
│   └── abis
    ___ typings
├── dev (include towns and river types, addresses)
│   ├── addresses
│   └── abis
    |__ typings
├── package.json
```