app: remove persisted react query store (#5610)
this is causing issues. i believe what is happening is, a query is
executed, stored in idb, then the query is garbage collected b/c we
don't have a long gcTime - it defaults to 5 mins - and so you go to load
a component that had this query previously, the data is not refetched
because of our other configurations (refetchOnMount, refetchOnFocus, etc
are typically off), the data is also gone from the cache, you get no
data and possibly unrecoverable errors that result in a black screen.

Case in point, you try to create a channel, exit the modal, come back
later, errors. Refreshes don't refetch either, I'm not sure why and I
would like to understand it better. I noticed this as well in another
part of the site I was developing on today, I kept loading a component
with outdated responses that did not reflect the updates I'd just made
in code.

Anyway, I do think we should have this, but some more investigation and
testing should be performed. I think we should use the hydrate callback
the persister provides to filter and store only specific queries. And in
that case we can just use localStorage instead of idb, just simpler imo.
I can't address this now so I'm just removing it.