app: swap out privy address for abstract account address in relevant spots (#5713)
We now have 2 user addresses we need to be aware of. 

The river `userId`, which is the privy address, which we need to use for
anything river related - signing messages, creating streams, etc.

Now there's the additional `abstractAccountAddress` which is primarily
for display purposes, but also key in several other areas:
- user gating on roles. This has been updated to pass the AA address.
- avatar image and bio storage. With this change, you'll have to
re-upload your avatar and bio
- user profiles. Profile slugs have been updated to be `/profile/<AA
address>`. The profile panel then does a lookup on the AA address to
grab the userId and load relevant data

I kept the privy wallet in the wallet panel for now, in case we all need
to grab it and transfer any funds