app: show authentication toast if privy is not authenticated (#5769)
As discussed in project design.

I witnessed Artem in a River authenticated but Privy unauthenticated
state. This is probably the random issue some of us have encountered
where there's no signer for on chain tx.

This state was previously impossible because we only loaded the
authenticated routes of the app if you were in an authenticated privy
state. However, we refactored for UX/performance so that app loading is
not reliant on privy status.

Which is all good until you run into an action that requires your privy
signer.

Unfortunately we cannot just re-log the user in to privy - there is
nothing to stop them from choosing to log in to privy with a different
account than the one they are already logged in to river with. Privy
just released a vanilla js sdk (which they encourage users not to use),
ultimately I'm hoping we can move over to that.

For now, if user runs into this, they are gonna have to go back through
the login flow. The plus side is, i've only heard of devs running into
this issue, and I suspect it has something to do with resetting site
data often.