Fix isOwner check in space creation flow (#5823)
There will be two triggers for `SpaceGettingStarted`. When we trigger
the first time, space is not available, then we check the owner on an
undefined space, and return false. It will automatically redirect to
`http://localhost:3000/t/undefined/threads`

We will navigate only when we check isOwner on an existing space id.