Fix for stuck membership in sdk (#94)
If the user stream doesn’t udpate, we need to retry the join or leave,
even if we have the correct membership in the channel. This is how we do
transactions in our distributed system, the group of writes isn’t
guaranteed to all succeed, but they should get to the right state on
retry.