HNT-5144 Update ts code to only do delegate signatures the one way (#57)
We only want to support one way to create deleagte signatures. We have
to support signing via a wallet, because that’s how privy works, so
that’s what we will support

This is a sdk breaking change for harmony, but not a data breaking
change, i’ll have to fix it up once it comes over

I did this change in two parts - the first commit moves some code from
the encryption package that didn’t need to be there. the second makes
the changes to the ts and go code