app: fix re-joining a town (#5971)
So I put this fix in for now to get back into gamma, but I'm not sure
that what the contract is doing is correct. In a nutshell, the
_isEntitled check for Permissions.JoinSpace returns true when you are
not a member, but false if you are a member.

```
    // if you're already a member and trying to join, return false
    if (
      isMember && permission == bytes32(abi.encodePacked(Permissions.JoinSpace))
    ) {
      return false;
    }
```

There is some UI in the app that blocks you from proceeding with joining
a town if your entitlement check fails. So now it checks for if you have
the permission, and if that fails, it performs an additional check on
the balance of your member NFT

This is the old _isEntitled code, it's hard for me to tell when this
changed since the repos were split, but it didn't used to have this
check in it on Permissions.JoinSpace.

It's confusing to have this entitlement check swap like this, is this
the correct behavior?

```
function _isEntitled(
    string memory channelId,
    address user,
    bytes32 permission
  ) internal view returns (bool entitled) {
    if (user == _owner()) return true;

    uint256 tokenId = MembershipStorage.layout().tokenIdByMember[user];
    if (_isBanned(tokenId) || _isBannedByChannel(channelId, tokenId))
      return false;

    EntitlementsManagerStorage.Layout storage ds = EntitlementsManagerStorage
      .layout();

    uint256 entitlementCount = ds.entitlements.length();

    for (uint256 i = 0; i < entitlementCount; i++) {
      if (
        IEntitlement(ds.entitlements.at(i)).isEntitled(
          channelId,
          user,
          permission
        )
      ) {
        entitled = true;
        break;
      }
    }
  }
```