app: refactor panels - part 1 (hnt-5273) (#6282)
The main focus of this PR is to enable right hand panel from just
anywhere. There's talks about introducing named outlets in react-router
just like in `Vue` however for the time being the best hack we've
figured is using searchParams.

- The main content is represented by its route like usual
`/t/town/channel`
- The side panel with the `?panel=` param

Main changes:
- I got rid of the `CentralPanelLayout`
- `usePanels()` returns a panel depending on the searchParam ans places
it right over the <Outlet /> in the main view
- `usePanelAction` returns a few utility actions such as `openPanel`,
`closePanel` etc.
- removed all duplicate routes and `createLink` for the latter (e.g.
/profile route)
- We handle closing, popping stacks internally from within the `<Panel
/>` so we don't have to do case-by-case callbacks. However I've kept and
renamed `onClose` - `onClosed` a callback triggered once a panel gets
closed. This is handy in a couple of places to save state or "mark as
viewed"


Last week I attempted to go a step further getting rid of all nested
routes to enable something like this on mobile. But it introduced so
many new issues (more than it solved) so I decided keeping it as
"simple" as possible for now while paving the way for the next
iteration. The main reason this is hard to achieve is that we rely on
the router to persist panels and would almost need to bypass it
completely (i.e use state).

```
<Home>
    <AnimatePresence>
        <Panel ?panel=home>
        <Panel ?panel=channel>
        <Panel ?panel=thread>
        ...
    </AnimatePresence>
</Home>
<Search>
    <AnimatePresence>
        <Panel ?panel=search>
        <Panel ?panel=profile>
        <Panel ?panel=dm>
        <Panel ?panel=dm-info>
        ...
    </AnimatePresence>
</Search>
...
```